+++ ava.config.js
@@ -1,0 +1,8 @@
+const config = {
+  verbose: true,
+  files: ['test/test-*'],
+  extensions: ['ts'],
+  require: ['esm', 'esbuild-register']
+};
+
+export default config;